Questions module done and category,quiz updated

+++ ExamPortal-Frontend/src/app/components/quiz/view-quiz/view-quiz.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { QuizService } from 'src/app/services/quiz.service';
+import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 
 @Component({
@@ -9,14 +10,14 @@ import Swal from 'sweetalert2';
 })
 export class ViewQuizComponent implements OnInit {
   quiz: any;
-  constructor(private quizService: QuizService) {}
+  constructor(private quizService: QuizService, private router: Router) {}
 
   ngOnInit(): void {
     this.reloadData();
   }
-  reloadData(){
-    this.quizService.getQuiz().subscribe(
-      (data: any) => {
+  reloadData() {
+    this.quizService.getAllQuiz().subscribe(
+      (data) => {
         this.quiz = data;
       },
       (error) => {
@@ -25,15 +26,28 @@ export class ViewQuizComponent implements OnInit {
       }
     );
   }
-  deleteQuiz(qid: any) {
-    this.quizService.deleteQuiz(qid).subscribe(
-      (data) => {
-        this.ngOnInit();
-        Swal.fire('Success!', 'Quiz Deleted ', 'success');
-      },
-      (error) => {
-        Swal.fire('Error!', 'server loading error', 'error');
+  deleteQuiz(qid:any) {
+    Swal.fire({
+      icon: 'info',
+      title: 'Are you sure you want to delete?',
+      confirmButtonText: 'Delete',
+      showCancelButton: true,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.quizService.deleteQuiz(qid).subscribe(
+          (data) => {
+            this.ngOnInit();
+            Swal.fire('Success!', 'Quiz Deleted ', 'success');
+          },
+          (error) => {
+            Swal.fire('Error!', 'server loading error', 'error');
+          }
+        );
       }
-    );
+    });
   }
+  updateQuiz(qid: any) {
+    this.router.navigate(['/admin/update-quiz/', qid]);
+  }
+  
 }
